+++ pages/admin/content-manager/[entity_type_slug]/index.js
@@ -1,18 +1,13 @@
 import CacheContext from "@/components/contexts/CacheContext";
-
 import React, { useEffect, useReducer, useRef, useContext, useState } from "react";
 import { slsFetch } from "@klaudsol/commons/lib/Client";
 import { useRouter } from "next/router";
-import { extractFiles } from "@/lib/s3FormController";
 import RootContext from '@/components/contexts/RootContext';
 
 /** kladusol CMS components */
 import AppCreatebutton from "@/components/klaudsolcms/buttons/AppCreateButton";
 import AppBackButton from "@/components/klaudsolcms/buttons/AppBackButton";
 import AppIconButton from "@/components/klaudsolcms/buttons/AppIconButton";
-import AppButtonLg from "@/components/klaudsolcms/buttons/AppButtonLg";
-import AppButtonSpinner from "@/components/klaudsolcms/AppButtonSpinner";
-import AppInfoModal from "@/components/klaudsolcms/modals/AppInfoModal";
 import AppContentManagerTable from "components/klaudsolcms/tables/AppContentManagerTable";
 import AppContentManagerTableIconView from "@/components/klaudsolcms/views/AppContentManagerIconView";
 import SkeletonTable from "components/klaudsolcms/skeleton/SkeletonTable";
@@ -20,7 +15,6 @@ import ContentManagerLayout from "components/layouts/ContentManagerLayout";
 
 /** react-icons */
 import { 
-  FaCheck,
   FaList,
   FaTh,
   FaDownload
@@ -32,7 +26,6 @@ import {
 } from "@/components/reducers/contentManagerReducer";
 import {
   LOADING,
-  SAVING,
   SET_ENTITY_TYPE_NAME,
   SET_COLUMNS,
   SET_VALUES,
@@ -42,8 +35,6 @@ import {
   SET_PAGE,
   PAGE_SETS_RENDERER,
   TOGGLE_VIEW,
-  SET_SHOW,
-  SET_MODAL_CONTENT,
   SET_DATA,
   SET_METADATA
 } from "@/lib/actions";
@@ -54,15 +45,8 @@ import { getSessionCache } from "@klaudsol/commons/lib/Session";
 import { useClientErrorHandler } from "@/components/hooks"
 import { handleDownloadCsv } from "@/lib/downloadCSV";
 import { Spinner } from "react-bootstrap";
-import { Formik, Form, Field } from "formik";
-import { sortByOrderAsc } from "@/components/Util";
-import AdminRenderer from "@/components/renderers/admin/AdminRenderer";
-import { uploadFilesToUrl } from "@/backend/data_access/S3";
-import GeneralHoverTooltip from "@/components/elements/tooltips/GeneralHoverTooltip";
-import { RiQuestionLine } from "react-icons/ri";
-import { slugTooltipText } from "@/constants";
-import classname from "classnames";
-import TypesValidator from "@/components/renderers/validation/RegexValidator";
+import { defaultEntityTypeVariant, entityTypeVariantsEnum } from "@/constants";
+import SingleType from "@/components/entity_types/SingleType";
 
 export default function ContentManager({ cache }) {
   const router = useRouter();
@@ -77,17 +61,15 @@ export default function ContentManager({ cache }) {
 
   const downloadCSVapi = `/api/downloadCsv?entity_type_slug=${entity_type_slug}`;
 
-  const [singleType, setSingleType] = useState(false);
+  const [variant, setVariant] = useState(defaultEntityTypeVariant);
   const [attributes, setAttributes] = useState({});
-  const formRef = useRef();
-
   const [entityTypeId, setEntityTypeId] = useState(0);
 
   /*** Entity Types List ***/
   useEffect(() => {
     (async () => {
       try {
-         dispatch({type:LOADING})
+        dispatch({ type: LOADING })
         if (controllerRef.current) {
           controllerRef.current.abort();
         }
@@ -106,7 +88,7 @@ export default function ContentManager({ cache }) {
         const attributes = values.metadata.attributes;
 
         setAttributes(attributes);
-        setSingleType(values.metadata.is_single_type);
+        setVariant(values.metadata.variant);
 
         dispatch({type: SET_DATA, payload: values.data});
         dispatch({type: SET_METADATA, payload: values.metadata});
@@ -130,7 +112,7 @@ export default function ContentManager({ cache }) {
         columns.unshift({ accessor: "slug", displayName: "SLUG" });
         columns.unshift({ accessor: "id", displayName: "ID" });
         dispatch({ type: SET_COLUMNS, payload: columns });
-        dispatch({ type: SET_VALUES, payload: values.metadata.is_single_type ? singleTypeEntries : entries });
+        dispatch({ type: SET_VALUES, payload: values.metadata.variant === entityTypeVariantsEnum.singleton ? singleTypeEntries : entries });
         controllerRef.current = null;
       } catch (ex) {
         errorHandler(ex);
@@ -151,85 +133,6 @@ export default function ContentManager({ cache }) {
     dispatch({ type: TOGGLE_VIEW, payload: view })
   }
 
-  const getFormikInitialVals = () => {
-    const { slug, id, ...initialValues } = state.values;
-    return state.values;
-  };
-
-  const onSubmit = (e) => {
-    e.preventDefault();
-    formRef.current.handleSubmit();
-    formRef.current.setTouched({ ...state.set_validate_all, slug: true });
-  };
-
-
-  const formatSlug = (slug) => {
-    return slug.toLowerCase().replace(/\s+/g, "-");
-  };
-
-  const formikParams = {
-    innerRef: formRef,
-    initialValues: getFormikInitialVals(),
-    onSubmit: (values) => {
-      (async () => {
-        try {
-          dispatch({ type: SAVING });
-          const { slug } = values;
-          const { data, fileNames, files } = await extractFiles(values);
-          // Checks if the entry is to be updated or not
-          // If there are no entries yet, it will call a different API (similar to create entry)
-          // Method is POST
-          const isUpdate = Object.keys(state.values).length > 0;
-          let url = `/api/${entity_type_slug}`;
-          let method = 'POST';
-          let formattedSlug = formatSlug(slug);
-          let entry = {
-            ...data,
-            fileNames,
-            slug: formattedSlug,
-            entity_type_id: entityTypeId,
-          }
-
-          // If there is already an existing entry, it will call a different API to update 
-          // Method is PUT
-          if (isUpdate) {
-            url = `/api/${entity_type_slug}/${values.id}`;
-            method = 'PUT';
-            entry = {
-              ...data,
-              fileNames,
-              entity_type_slug,
-              entity_id: values.id,
-            }
-          }
-        
-  
-          const response = await slsFetch(url, {
-            method: method,
-            headers: {
-              "Content-type": "application/json",
-            },
-            body: JSON.stringify(entry),
-          });
-
-          const { message, presignedUrls } = await response.json();
-
-          if (files.length > 0) await uploadFilesToUrl(files, presignedUrls);
-
-          dispatch({
-            type: SET_MODAL_CONTENT,
-            payload: message,
-          });
-          dispatch({ type: SET_SHOW, payload: true });
-        } catch (ex) {
-          errorHandler(ex);
-        } finally {
-          dispatch({ type: CLEANUP });
-        }
-      })();
-    },
-  };
-
   return (
     <CacheContext.Provider value={cache}>
       <div className="d-flex flex-row mt-0 pt-0 mx-0 px-0">
@@ -249,7 +152,7 @@ export default function ContentManager({ cache }) {
                 <p> {state.values.length} entries found </p>
               </div>
               <div className="general-row-center" style={{ gap: '5px'}}>
-              {capabilities.includes(writeContents) && !singleType &&
+              {capabilities.includes(writeContents) && variant === entityTypeVariantsEnum.collection &&
               <AppCreatebutton
                 link={`/admin/content-manager/${entity_type_slug}/create`}
                 title="Create new entry"
@@ -269,6 +172,7 @@ export default function ContentManager({ cache }) {
               </div>
             </div>
             <div className="d-flex justify-content-between align-items-center px-0 mx-0 pb-3">
+            
               {/*<div className="d-flex flex-row px-0">
                 TODO:
                   <AppIconButton icon={<FaSearch/>} /> 
@@ -281,7 +185,7 @@ export default function ContentManager({ cache }) {
                   id="dropdown_general"
                   isCheckbox={true}
               />*/}
-              {!singleType && 
+              {variant === entityTypeVariantsEnum.collection && 
               <div className="general-row-end" style={{ gap: '5px '}}>
                 <AppIconButton icon={<FaList/>} selected={state.view ==='list'} onClick={() => handleView('list')}/>
                 <AppIconButton icon={<FaTh />} selected={state.view === 'icon'} onClick={() => handleView('icon')}/>
@@ -289,87 +193,40 @@ export default function ContentManager({ cache }) {
               {/* <AppIconButton icon={<BsGearFill/>} />  */}
             </div>
 
-            {(state.isLoading && state.firstFetch) && <SkeletonTable />}
-            {!state.isLoading && singleType && (
-                    <Formik {...formikParams}>
-                      {(props) => (
-                        <Form>
-                          {Object.keys(state.values).length === 0 && 
-                          <>
-                            <div className="d-flex flex-row mx-0 my-0 px-0 py-0"> 
-                          <p className="general-input-title-slug"> Slug </p> 
-                          <GeneralHoverTooltip 
-                            icon={<RiQuestionLine className="general-input-title-slug-icon"/>}
-                            className="general-table-header-slug"
-                            tooltipText={slugTooltipText}
-                            position="left"
-                          /> 
-                          </div>
-                          <Field
-                            name="slug"
-                            validate={(e) => TypesValidator(e, "text")}
-                          >
-                            {({ field, meta }) => (
-                              <div>
-                                <input
-                                  type="text"
-                                  {...field}
-                                  className={classname("general-input-text", {"general-input-error" : meta.touched && meta.error})}
-                                  style={
-                                    meta.touched && meta.error
-                                      ? {
-                                          borderColor: "red",
-                                          outlineColor: "red",
-                                        }
-                                      : {}
-                                  }
-                                />
-                                {meta.touched && meta.error && (
-                                  <div className="general-input-error-text">
-                                    {meta.error}
-                                  </div>
-                                )}
-                              </div>
-                            )}
-                          </Field>
-                          </>}
-                          {Object.entries(attributes)
-                            .sort(sortByOrderAsc)
-                            .map(([attributeName, attribute]) => {
-                              return (
-                                <div key={attributeName}>
-                                  <p className="general-input-title"> {attributeName.replaceAll('_', " ")}  </p>
-                                  <AdminRenderer
-                                    errors={props.errors}
-                                    touched={props.touched}
-                                    type={attribute.type}
-                                    name={attributeName}
-                                    disabled={!capabilities.includes(writeContents)}
-                                  />
-                                </div>
-                              );
-                            })}
-                        </Form>
-                      )}
-                    </Formik>
-                  )}
-            {(state.firstFetch ? !state.isLoading : !state.firstFetch) && state.view === 'list' && !singleType &&  (
-              <AppContentManagerTable
-                columns={state.columns}
+            {variant === entityTypeVariantsEnum.singleton && 
+              <SingleType 
+                loading={state.isLoading}
                 entries={state.values}
                 entity_type_slug={entity_type_slug}
-                data={state.data}
-                metadata={state.metadata}
-              />
+                attributes={attributes}
+                capabilities={capabilities}
+                entity_type_id={entityTypeId}
+              />}
+
+            {(state.isLoading && state.firstFetch) && <SkeletonTable />}
+      
+            {(state.firstFetch ? !state.isLoading : !state.firstFetch) && 
+              state.view === 'list' &&
+              variant === entityTypeVariantsEnum.collection && (
+                <AppContentManagerTable
+                  columns={state.columns}
+                  entries={state.values}
+                  entity_type_slug={entity_type_slug}
+                  data={state.data}
+                  metadata={state.metadata}
+                />
             )}
-            {(state.firstFetch ? !state.isLoading : !state.firstFetch) && state.view === 'icon' && !singleType && (
+            {(state.firstFetch ? !state.isLoading : !state.firstFetch) && 
+              state.view === 'icon' && 
+              variant === entityTypeVariantsEnum.collection && (
               <AppContentManagerTableIconView
                 columns={state.columns}
                 entries={state.values}
                 entity_type_slug={entity_type_slug}
               />
             )}
-            {(state.firstFetch ? !state.isLoading : !state.firstFetch) && !singleType && (
+            {(state.firstFetch ? !state.isLoading : !state.firstFetch) && 
+            variant === entityTypeVariantsEnum.collection && (
               <AppContentPagination
                 dispatch={dispatch}
                 defaultEntry={state.entry}
@@ -384,21 +241,7 @@ export default function ContentManager({ cache }) {
               />
             )}
           <div className="py-3" />
-          {!state.isLoading && singleType && 
-            <div className="d-flex flex-row justify-content-center">
-              {capabilities.includes(writeContents) && <>
-              <AppButtonLg
-                title="Cancel"
-                onClick={!state.isSaving ? () => router.push(`/admin/content-manager/${entity_type_slug}`) : null}
-                className="general-button-cancel mb-3"
-              />
-              <AppButtonLg
-                title={state.isSaving ? "Saving" : "Save"}
-                icon={state.isSaving ? <AppButtonSpinner /> : <FaCheck className="general-button-icon"/>}
-                onClick={!state.isSaving ? onSubmit : null}
-                className="general-button-save mb-3"
-              /></>}
-            </div>}
+
           </div>
 
           {/*<div className="d-flex justify-content-between align-items-center">
@@ -413,18 +256,7 @@ export default function ContentManager({ cache }) {
             <button className="btn_arrows"> <FaChevronRight className="mb-2 ml-1" style={{fontSize: "10px"}}/> </button>
           </div>
         </div>*/}
-                  <AppInfoModal
-            show={state.show}
-            onClose={() => (
-              dispatch({ type: SET_SHOW, payload: false }),
-              router.reload()
-            )}
-            modalTitle="Success"
-            buttonTitle="Close"
-          >
-            {" "}
-            {state.modalContent}{" "}
-          </AppInfoModal>
+                  
         </ContentManagerLayout>
       </div>
     </CacheContext.Provider>
